The Phoenix Sepsis Score

The diagnostic Phoenix Spesis Criteria based on four organ dysfunction scores, respiratory, cardiovascular, coagulation, and neurologic. A score of 2 or more indicates sepsis.

phoenix(
  vasoactives,
  lactate,
  age,
  map,
  pf_ratio,
  sf_ratio,
  imv,
  other_respiratory_support,
  platelets,
  inr,
  d_dimer,
  fibrinogen,
  gcs,
  fixed_pupils,
  data,
  ...
)

Arguments

vasoactives

an integer vector, the number of systemic vasoactive medications being administered to the patient

lactate

numeric vector with the lactate value in mmol/L

age

numeric vector age in months

map

numeric vector, mean arterial pressure in mmHg

pf_ratio

numeric vector

sf_ratio

numeric vector

imv

invasive mechanical ventilation; numeric or integer vector, (0 = not intubated; 1 = intubated)

other_respiratory_support

other respiratory support; numeric or integer vector, (0 = no support; 1 = support)

platelets

numeric vector for platelets counts in units of 1,000/uL

inr

numeric vector for the internation normlised ratio blood test

d_dimer

numeric vector for D-Dimer, units of mg/L FEU

fibrinogen

numeric vector units of mg/dL

gcs

integer vector; total Glasgow Comma Score

fixed_pupils

integer vector; 1 = bilaterally fixed pupil, 0 = otherwise

data

a list, data.frame, or environment containing the input vectors

...

pass through

Value

A data.frame with five columns:

phoenix_respiratory_score

phoenix_cardiovascular_score

phoenix_coagulation_score

phoenix_neurologic_score

phoenix_sepsis_score

As with all other Phoenix oragan system scores, missing values in the data set will map to a score of zero - this is consistent with the development of the criteria.

Details

The details of each of the four component scores are found in there respective help files.

References

Sanchez-Pinto LN, Bennett TD, DeWitt PE, et al. Development and Validation of the Phoenix Criteria for Pediatric Sepsis and Septic Shock. JAMA. Published online January 21, 2024. doi:10.1001/jama.2024.0196

Schlapbach LJ, Watson RS, Sorce LR, et al. International Consensus Criteria for Pediatric Sepsis and Septic Shock. JAMA. Published online January 21, 2024. doi:10.1001/jama.2024.0179

See also

Examples

DF <-
  expand.grid(
    # coagulation varaiables
    plts = c(20, 100, 150),
    inr  = c(0.2, 1.3, 1.8),
    ddmr = c(1.7, 2.0, 2.8),
    fib  = c(88, 100, 120),

    # cardiovascular variabls
    vasos = c(0:3),
    lactate = c(3.2, 7.8, 14),
    age = c(0.4, 3, 18, 45, 61, 164),
    map = c(16, 33, 55),

    # neurologic variables
    gcs = c(8, 12),
    pupils = c(0, 1),

    # respiratory variables
    pfr = c(500, 350, 187, 56),
    sfr = c(300, 254, 177, 76),
    vent = c(0, 1),
    o2  = c(0, 1)
    )

PSS <-
  phoenix(
    vasoactives = vasos,
    lactate = lactate,
    age = age,
    map = map,
    pf_ratio = pfr,
    sf_ratio = sfr,
    imv = vent,
    other_respiratory_support = o2,
    platelets = plts,
    inr = inr,
    d_dimer = ddmr,
    fibrinogen = fib,
    gcs = gcs,
    fixed_pupils = pupils,
    data = DF
  )
#> Error in eval(as.call(cl)): object 'DF' not found

head(PSS)
#> Error in eval(expr, envir, enclos): object 'PSS' not found