Add TopicScoreMatrix to ArrowFiles or an ArchRProject

This function, for each sample, will independently compute sums of score for each tile per cell and then infer gene topic scores. This function is adapted from addGeneScoreMatrix function from the ArchR package.

Usage

scoreGeneByPeak(
  genes = NULL,
  peaks = NULL,
  scoreMat = NULL,
  geneModel = "exp(-abs(x)/5000) + exp(-1)",
  peakWidth = 500,
  method = "sum",
  extendUpstream = c(1000, 1e+05),
  extendDownstream = c(1000, 1e+05),
  geneUpstream = 5000,
  geneDownstream = 0,
  useGeneBoundaries = TRUE,
  useTSS = FALSE,
  extendTSS = FALSE,
  geneScaleFactor = 5,
  excludeChr = c("chrY", "chrM"),
  blacklist = NULL,
  subThreads = 1,
  tstart = NULL,
  logFile = NULL
)

Arguments

genes

A stranded GRanges object containing the ranges associated with all gene start and end coordinates.

peaks

A GRanges object containing the peaks with a score column to calculate gene-level scores.

scoreMat

geneModel

A string giving a "gene model function" used for weighting peaks for gene score calculation. This string should be a function of x, where x is the stranded distance from the transcription start site of the gene.

extendUpstream

The minimum and maximum number of basepairs upstream of the transcription start site to consider for gene activity score calculation.

extendDownstream

The minimum and maximum number of basepairs downstream of the transcription start site or transcription termination site (based on 'useTSS') to consider for gene activity score calculation.

geneUpstream

An integer describing the number of bp upstream the gene to extend the gene body. This effectively makes the gene body larger as there are proximal peaks that should be weighted equally to the gene body. This parameter is used if 'useTSS=FALSE'.

geneDownstream

An integer describing the number of bp downstream the gene to extend the gene body.This effectively makes the gene body larger as there are proximal peaks that should be weighted equally to the gene body. This parameter is used if 'useTSS=FALSE'.

useGeneBoundaries

A boolean value indicating whether gene boundaries should be employed during gene activity score calculation. Gene boundaries refers to the process of preventing tiles from contributing to the gene score of a given gene if there is a second gene's transcription start site between the tile and the gene of interest.

useTSS

A boolean describing whether to build gene model based on gene TSS or the gene body.

extendTSS

A boolean describing whether to extend the gene TSS. By default useTSS uses the 1bp TSS while this parameter enables the extension of this region with 'geneUpstream' and 'geneDownstream' respectively.

geneScaleFactor

A numeric scaling factor to weight genes based on the inverse of there length i.e. (Scale Factor)/(Gene Length). This is scaled from 1 to the scale factor. Small genes will be the scale factor while extremely large genes will be closer to 1. This scaling helps with the relative gene score value.

excludeChr

A character vector containing the seqnames of the chromosomes that should be excluded from this analysis.

blacklist

A GRanges object containing genomic regions to blacklist that may be extremeley over-represented and thus biasing the geneScores for genes nearby that locus.

subThreads

tstart

logFile

The path to a file to be used for logging ArchR output.

matrixName

The name to be used for storage of the gene activity score matrix in the provided ArchRProject or ArrowFiles.

force

A boolean value indicating whether to force the matrix indicated by matrixName to be overwritten if it already exist in the given input.