Phoenix Sepsis Neurological Score

Assessment of neurologic dysfunction based on Glasgow Coma Scale (GCS) and pupil reactivity. This score is part of the diagnostic Phoenix Sepsis criteria and Phoenix 8 Sepsis criteria.

phoenix_neurologic(gcs, fixed_pupils, data = parent.frame(), ...)

Arguments

gcs

integer vector; total Glasgow Comma Score

fixed_pupils

integer vector; 1 = bilaterally fixed pupil, 0 = otherwise

data

a list, data.frame, or environment containing the input vectors

...

pass through

Value

an integer vector with values 0, 1, or 2.

Details

Scoring:

Bilaterally fixed pupil	2 points

Glasgow Coma Score (total) less or equal 10	1 point

Reactive pupils and GCS > 10	0 point

Missing value will be mapped to a value of 0 as was done when developing the Phoenix criteria. Note that this is done on a input by input basis. That is, if pupil reactivity is missing but GCS (total) = 9, then the score is one point.

GCS total is the sum of a score based on eyes, motor control, and verbal responsiveness.

Eye response:

no eye opening,

eye opening to pain,

eye opening to sound,

eyes open spontaneously.

Verbal response:

no verbal response,

incomprehensible sounds,

inappropriate words,

confused,

orientated

Motor response:

no motor response,

abnormal extension to pain,

abnormal flexion to pain,

withdrawal from pain,

localized pain,

obeys commands

References

Sanchez-Pinto LN, Bennett TD, DeWitt PE, et al. Development and Validation of the Phoenix Criteria for Pediatric Sepsis and Septic Shock. JAMA. Published online January 21, 2024. doi:10.1001/jama.2024.0196

Schlapbach LJ, Watson RS, Sorce LR, et al. International Consensus Criteria for Pediatric Sepsis and Septic Shock. JAMA. Published online January 21, 2024. doi:10.1001/jama.2024.0179

See also

Examples

DF <- expand.grid(gcs = c(3:15, NA), pupils = c(0, 1, NA))
DF$target <- 0L
DF$target[DF$gcs <= 10] <- 1L
DF$target[DF$pupils == 1] <- 2L
DF$current <- phoenix_neurologic(gcs, pupils, DF)
stopifnot(identical(DF$target, DF$current))
DF
#>    gcs pupils target current
#> 1    3      0      1       1
#> 2    4      0      1       1
#> 3    5      0      1       1
#> 4    6      0      1       1
#> 5    7      0      1       1
#> 6    8      0      1       1
#> 7    9      0      1       1
#> 8   10      0      1       1
#> 9   11      0      0       0
#> 10  12      0      0       0
#> 11  13      0      0       0
#> 12  14      0      0       0
#> 13  15      0      0       0
#> 14  NA      0      0       0
#> 15   3      1      2       2
#> 16   4      1      2       2
#> 17   5      1      2       2
#> 18   6      1      2       2
#> 19   7      1      2       2
#> 20   8      1      2       2
#> 21   9      1      2       2
#> 22  10      1      2       2
#> 23  11      1      2       2
#> 24  12      1      2       2
#> 25  13      1      2       2
#> 26  14      1      2       2
#> 27  15      1      2       2
#> 28  NA      1      2       2
#> 29   3     NA      1       1
#> 30   4     NA      1       1
#> 31   5     NA      1       1
#> 32   6     NA      1       1
#> 33   7     NA      1       1
#> 34   8     NA      1       1
#> 35   9     NA      1       1
#> 36  10     NA      1       1
#> 37  11     NA      0       0
#> 38  12     NA      0       0
#> 39  13     NA      0       0
#> 40  14     NA      0       0
#> 41  15     NA      0       0
#> 42  NA     NA      0       0