Phoenix Renal Score

Phoenix Renal Score

phoenix_renal(creatinine, age, data = parent.frame(), ...)

Arguments

creatinine

numeric vector; units of mg/dL

age

numeric vector age in months

data

a list, data.frame, or environment containing the input vectors

...

pass through

Value

a integer vector with values 0, 1, or 2

As with all other Phoenix organ system scores, missing values in the data set will map to a score of zero - this is consistent with the development of the criteria.

References

Sanchez-Pinto LN, Bennett TD, DeWitt PE, et al. Development and Validation of the Phoenix Criteria for Pediatric Sepsis and Septic Shock. JAMA. Published online January 21, 2024. doi:10.1001/jama.2024.0196

Schlapbach LJ, Watson RS, Sorce LR, et al. International Consensus Criteria for Pediatric Sepsis and Septic Shock. JAMA. Published online January 21, 2024. doi:10.1001/jama.2024.0179

See also

Examples

DF <- expand.grid(age = c(NA, 0.4, 1, 3, 12, 18, 24, 45, 60, 61, 144, 145),
                  creatinine = c(NA, seq(0.0, 1.1, by = 0.1)))
DF$card <- phoenix_renal(age = age, creatinine = creatinine, data = DF)

head(DF)
#>    age creatinine card
#> 1   NA         NA    0
#> 2  0.4         NA    0
#> 3  1.0         NA    0
#> 4  3.0         NA    0
#> 5 12.0         NA    0
#> 6 18.0         NA    0